+++ Classes/iOctocat.m
@@ -6,6 +6,7 @@
 #import "GHUser.h"
 #import "GHOrganization.h"
 #import "GHOrganizations.h"
+#import "GHNotifications.h"
 #import "NSString+Extensions.h"
 #import "NSDictionary+Extensions.h"
 #import "MenuController.h"
@@ -17,6 +18,7 @@
 #import "NSDate+Nibware.h"
 
 #define kClearAvatarCacheDefaultsKey @"clearAvatarCache"
+#define kUserNotificationsCountKeyPath @"user.notifications.notificationsCount"
 
 
 @interface iOctocat () <UIApplicationDelegate, BITHockeyManagerDelegate, BITCrashManagerDelegate, BITUpdateManagerDelegate>
@@ -57,7 +59,11 @@
 
 - (void)setCurrentAccount:(GHAccount *)account {
 	[self clearUserObjectCache];
+	[_currentAccount removeObserver:self forKeyPath:kUserNotificationsCountKeyPath];
 	_currentAccount = account;
+	[_currentAccount addObserver:self forKeyPath:kUserNotificationsCountKeyPath options:NSKeyValueObservingOptionNew context:nil];
+	NSInteger badgeNumber = self.currentAccount.user.notifications.notificationsCount;
+	[[UIApplication sharedApplication] setApplicationIconBadgeNumber:badgeNumber];
 	if (!self.currentAccount) {
 		UIBarButtonItem *btnItem = self.menuNavController.topViewController.navigationItem.rightBarButtonItem;
 		self.menuNavController.topViewController.navigationItem.rightBarButtonItem = nil;
@@ -196,6 +202,9 @@
 		if (user.gravatar) {
 			[IOCAvatarCache cacheGravatar:user.gravatar forIdentifier:user.login];
 		}
+	} else if ([keyPath isEqualToString:kUserNotificationsCountKeyPath]) {
+		NSInteger badgeNumber = [change safeIntegerForKey:@"new"];
+		[[UIApplication sharedApplication] setApplicationIconBadgeNumber:badgeNumber];
 	}
 }